+++ bukkit/src/main/java/de/maxhenkel/voicechat/config/ServerConfig.java
@@ -20,6 +20,8 @@ public class ServerConfig {
     public final ConfigEntry<Boolean> allowRecording;
     public final ConfigEntry<Boolean> spectatorInteraction;
     public ConfigEntry<Boolean> spectatorPlayerPossession;
+    public ConfigEntry<Boolean> forceVoiceChat;
+    public ConfigEntry<Integer> loginTimeout;
 
     public ServerConfig(ConfigBuilder builder) {
         voiceChatPort = builder.integerEntry("port", 24454, -1, 65535);
@@ -37,6 +39,8 @@ public class ServerConfig {
         allowRecording = builder.booleanEntry("allow_recording", true);
         spectatorInteraction = builder.booleanEntry("spectator_interaction", false);
         spectatorPlayerPossession = builder.booleanEntry("spectator_player_possession", false);
+        forceVoiceChat = builder.booleanEntry("force_voice_chat", false);
+        loginTimeout = builder.integerEntry("login_timeout", 10_000, 100, Integer.MAX_VALUE);
     }
 
     public enum Codec {