+++ forge/src/main/java/de/maxhenkel/voicechat/config/ForgeServerConfig.java
@@ -87,6 +87,16 @@ public class ForgeServerConfig extends ServerConfig {
                 .comment("If spectators can talk to players they are spectating")
                 .define("voice_chat.spectator_player_possession", false)
         );
+        forceVoiceChat = wrapConfigEntry(builder
+                .worldRestart()
+                .comment("If players without the mod should get kicked from the server")
+                .define("voice_chat.force_voice_chat", false)
+        );
+        loginTimeout = wrapConfigEntry(builder
+                .worldRestart()
+                .comment("The amount of milliseconds, the server should wait to check if the player has the mod installed", "Only active when force_voice_chat is set to true")
+                .defineInRange("voice_chat.login_timeout", 10_000, 100, Integer.MAX_VALUE)
+        );
     }
 
     public static <T> ConfigEntry<T> wrapConfigEntry(ForgeConfigSpec.ConfigValue<T> configValue) {